+++ src/bika/lims/browser/analyses/view.py
@@ -914,7 +914,7 @@ class AnalysesView(ListingView):
             val = json.loads(value)
             if isinstance(val, (list, tuple, set)):
                 value = val
-        except ValueError:
+        except (ValueError, TypeError):
             pass
         if not isinstance(value, (list, tuple, set)):
             value = [value]