+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 2.3.0 (unreleased)
 ------------------
 
+- #2011 Removed dependency to Products.TextIndexNG3
+
 
 
 2.2.0 (2022-06-11)