+++ CHANGES.rst
@@ -6,7 +6,8 @@ Changelog
 ------------------
 
 - #2012 Remove stale supply order code
-
+- #2011 Removed dependency to Products.TextIndexNG3
+- #2010 Fix TypeError in Analyses Listing View
 
 
 2.2.0 (2022-06-11)